Added function to get the user authenticated via jwt token (GetCurrentUser) and create a RequestInitiatorDTO, WIP usage of the authorization.

+++ Backend/TrelloClone/Controllers/UserController.cs
@@ -1,4 +1,9 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.IdentityModel.Tokens;
+using Newtonsoft.Json;
+using System.Security.Claims;
+using System.Text;
 using TrelloClone.DTO.Creation;
 using TrelloClone.DTO.Display;
 using TrelloClone.Exceptions;
@@ -10,6 +15,7 @@ namespace TrelloClone.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+    //[Authorize]
     public class UserController : Controller
     {
         private readonly IUserService _userService;
@@ -23,6 +29,9 @@ namespace TrelloClone.Controllers
         [ProducesResponseType(200, Type = typeof(IEnumerable<UserDTO>))]
         public async Task<IActionResult> GetUsers()
         {
+
+            var initiatingUser = GetCurrentUser();
+
             var users = await _userService.GetUsers();
 
             if (!ModelState.IsValid)
@@ -57,6 +66,7 @@ namespace TrelloClone.Controllers
         [Consumes("application/json")]
         [ProducesResponseType(200, Type = typeof(UserDTO))]
         [TrelloControllerFilter]
+        [AllowAnonymous]
         public async Task<IActionResult> RegisterUser(CredentialUserDTO newUser)
         {
             var createdUser = await _userService.CreateUser(newUser);
@@ -68,9 +78,22 @@ namespace TrelloClone.Controllers
 
 
         [HttpDelete("{username}")]
+        [Authorize(Roles = "Admin")]
         public async Task<IActionResult> DeleteUser(string username)
         {
-            await _userService.DeleteUser(username);
+            var requestInitiatingUser = GetCurrentUser();
+            if (requestInitiatingUser != null)
+            {
+                /*if (requestInitiatingUser.Username == username)
+                {*/
+                await _userService.DeleteUser(username);
+                //}
+                /*else
+                {
+                    return Unauthorized("Insufficient permissions to perform this action.");
+                }*/
+            }
+
 
             return Ok();
         }
@@ -84,6 +107,19 @@ namespace TrelloClone.Controllers
             return Ok(user);
         }
 
+        /*[HttpPost("/login")]
+        [Consumes("application/json")]
+        [ProducesResponseType(200, Type = typeof(string))]
+        [TrelloControllerFilter]
+        public async Task<IActionResult> Login(CredentialUserDTO userLogin)
+        {
+            var token = await _userService.Login(userLogin);
+
+            Console.WriteLine("TEST");
+            return Ok(token);
+        }
+*/
+
         /*[HttpGet("/TEST")]
         public async Task<IActionResult> TestFunction(string username)
         {
@@ -91,5 +127,26 @@ namespace TrelloClone.Controllers
 
             return Ok(user);
         }*/
+
+        private RequestInitiatorDTO GetCurrentUser()
+        {
+            var identity = HttpContext.User.Identity as ClaimsIdentity;
+
+            if (identity != null)
+            {
+                var userClaims = identity.Claims;
+
+                return new RequestInitiatorDTO
+                {
+                    Username = userClaims.FirstOrDefault(o => o.Type == ClaimTypes.NameIdentifier)?.Value,
+                    Role = userClaims.FirstOrDefault(o => o.Type == ClaimTypes.Role)?.Value,
+                    BoardMemberships = JsonConvert
+                        .DeserializeObject<int[]>(userClaims
+                                                    .FirstOrDefault(o => o.Type == Helper.Helper.authorizedBoardsClaimName)?.Value)
+
+                };
+            }
+            return null;
+        }
     }
 }
